rdpmKeyBlocker
rdpm (RustDesk Privacy Mode) is a feature in RustDesk which prevents a user from seeing their screen. This is followed by text informing them they are in Privacy Mode and they can undo  the mode with a keyboard combination. This tool blocks input for this combination preventing them from undoing the privacy screen